Move test_load_crop_metadata to testing/api/ (unit test, not CLI)
